+++ be/promotion/src/main/java/woowa/promotion/admin/admin/application/dto/request/SignupServiceRequest.java
@@ -1,4 +1,4 @@
-package woowa.promotion.admin.admin.application.dto;
+package woowa.promotion.admin.admin.application.dto.request;
 
 import woowa.promotion.admin.admin.presentation.dto.request.SignupRequest;
