+++ be/promotion/src/test/java/woowa/promotion/admin/admin/application/AuthServiceTest.java
@@ -1,16 +1,16 @@
 package woowa.promotion.admin.admin.application;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import woowa.promotion.admin.admin.application.dto.SignupServiceRequest;
+import woowa.promotion.admin.admin.application.dto.request.SignupServiceRequest;
 import woowa.promotion.admin.admin.domain.Admin;
 import woowa.promotion.fixture.FixtureFactory;
 import woowa.promotion.global.security.hash.PasswordEncoder;
 import woowa.promotion.util.ApplicationTest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DisplayName("[비즈니스 로직 테스트] 관리자")
 class AuthServiceTest extends ApplicationTest {
